Raise exception if XSLT processing required
Currently XSLT processing is not supported in .NET Core (see issue #3).

If a TriX source files contains an XSLT stylesheet PI, the chances are
extremely high that the results of parsing the XML without applying the
XSLT transform will be invalid. For this reason we raise an
RdfParseException rather than just a warning.