Skip test for known issue
This test requires that the parser rejects IRIs that are not in Unicode
Normal Form C. .NET Core does not support the test for string
normalization (see #1) and as a result we cannot pass this test.

Renaming the file extension to .skip prevents the RDF/XML parser suites
from attempting to run this test. We should instead document this as a
known issue.